feat(middleware): add CORS options and refactor utility functions
- Added new CORS-related route options (WithCORS, WithCustomCORS, WithTUSCORS) in middleware/option/package
- Consolidated duplicate test file by moving route_builder_options_test.go to middleware/option/options_test.go
- Simplified Wrap function in util/convert/convert.go to use echo.WrapMiddleware
- Removed redundant TestWrap tests since functionality is now delegated to echo
- Deleted obsolete route_builder_options.go file
- Updated go.mod dependencies (gswagger and portal-router versions)
