+++ util/convert/convert.go
@@ -9,19 +9,7 @@ import (
 // Wrap converts a standard http.Handler middleware to an echo.MiddlewareFunc.
 // It properly propagates errors from the echo handler chain.
 func Wrap(mw func(http.Handler) http.Handler) echo.MiddlewareFunc {
-	return func(next echo.HandlerFunc) echo.HandlerFunc {
-		return func(c echo.Context) error {
-			var handlerErr error
-			
-			handler := mw(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				c.SetRequest(r)
-				handlerErr = next(c)
-			}))
-			
-			handler.ServeHTTP(c.Response(), c.Request())
-			return handlerErr
-		}
-	}
+	return echo.WrapMiddleware(mw)
 }
 
 // Unwrap converts an echo.MiddlewareFunc to a standard http.Handler middleware.
@@ -30,12 +18,12 @@ func Unwrap(mw echo.MiddlewareFunc) func(http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			e := echo.New()
 			c := e.NewContext(r, w)
-			
+
 			handler := mw(func(c echo.Context) error {
 				next.ServeHTTP(c.Response(), c.Request())
 				return nil
 			})
-			
+
 			// Execute the middleware chain and handle any errors
 			if err := handler(c); err != nil {
 				e.HTTPErrorHandler(err, c)
